+++ Smartstore.AI.GeminiClient/changelog.md
@@ -1,7 +1,12 @@
 # Release Notes
 
-## Smartstore.AI.GeminiClient 1.0.1
+## Smartstore.AI.GeminiClient 1.0.2
+### Improvements
 
+- Enable to pass file URLs to `DeleteFileAsync` method.
+
+
+## Smartstore.AI.GeminiClient 1.0.1
 ### New Features
 
 - Get a list of available Gemini models.